Removing Card to Card PairingIdx
Removing this as I don't think it will be especially helpful for the terminal to have a pairingIdx by which to refer to the card to card secure channel. It would be better to have the card only have one card to session available at a time, and simply to default to this one. If a more permanent reference to the C2C channel is needed, using the counterparty card's pubKey as the reference would be more robust to manage across time, since it does not require state to be held on a single terminal to refer to an old channel.